# src/docs/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Split API",
    "description": "API for split",
    "version": "1.0.0"
  },
  "host": [
    "api.split-app.fr"
  ],
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Enter your Bearer token in the format **Bearer &lt;token&gt;**"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/user/me": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "Get all information of the connected user.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "User information retrieved successfully.",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "401": {
            "description": "Unauthorized. No valid token provided.",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 401
                },
                "msg": {
                  "type": "string",
                  "example": "Unauthorized."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/user/": {
      "delete": {
        "tags": [
          "User"
        ],
        "description": "Delete an User.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "User information deleted successfully.",
            "schema": {
              "type": "object",
              "properties": {
                "msg": {
                  "type": "string",
                  "example": ""
                },
                "status": {
                  "type": "number",
                  "example": 200
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Unauthorized. No valid token provided.",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 401
                },
                "msg": {
                  "type": "string",
                  "example": "Unauthorized."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/user/register": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "Register a new user.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "User registration data",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/user/update": {
      "put": {
        "tags": [
          "User"
        ],
        "description": "Update user details.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "User update data",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User updated successfully.",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "User not found.",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 404
                },
                "msg": {
                  "type": "string",
                  "example": "User not found."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "422": {
            "description": "Unprocessable entity.",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 422
                },
                "msg": {
                  "type": "string",
                  "example": "Required fields missing."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/user/image": {
      "delete": {
        "tags": [
          "User"
        ],
        "description": "Delete user image.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "User updated successfully.",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "User not found.",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 404
                },
                "msg": {
                  "type": "string",
                  "example": "User not found."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "422": {
            "description": "Unprocessable entity.",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 422
                },
                "msg": {
                  "type": "string",
                  "example": "Required fields missing."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/user/check/username/{username}": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "Check if username already use.",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "username to check"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/user/login": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "User login.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "User login data",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "user@example.com"
                },
                "password": {
                  "type": "string",
                  "example": "Password123"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User login successful.",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    },
...
    "/user/refresh-token/{refreshToken}": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "Refresh user token.",
        "parameters": [
          {
            "name": "refreshToken",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Refresh token"
          }
        ],
        "responses": {
          "200": {
            "description": "User token refreshed successfully.",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "401": {
            "description": "Unauthorized. No valid token provided.",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 401
                },
                "msg": {
                  "type": "string",
                  "example": "Aucun token valide trouvé."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    }
  },
  "definitions": {
...
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "firstName": {
          "type": "string",
          "example": "John"
        },
        "lastName": {
          "type": "string",
          "example": "Doe"
        },
        "username": {
          "type": "string",
          "example": "johndoe"
        },
        "email": {
          "type": "string",
          "example": "john.doe@gmail.com"
        },
        "isCompleted": {
          "type": "boolean",
          "example": true
        },
        "imageLink": {
          "type": "string",
          "example": "https://example.com/image.jpg"
        },
        "isGuest": {
          "type": "boolean",
          "example": false
        }
      }
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ]
}